feat: center point as cavas center

+++ library/src/common.rs
@@ -3,7 +3,7 @@ pub type Point = (i32, i32);
 pub type Image = Vec<Vec<Rgba>>;
 pub type Edge = (Point, Point);
 //3d point in homogeneous coordinates
-pub type HomogeneousPoint = (f32, f32, f32, f32);
+pub type HomogeneousPoint = (f64, f64, f64, f64);
 // pub type Border = (i32, i32, i32, i32);
 pub type HomogeneousEdge = (HomogeneousPoint, HomogeneousPoint);
 pub struct Border {
